requirements/base.txt
redis==4.5.4
elasticsearch[async]==8.7.0
fastapi==0.95.1
orjson==3.8.10
pydantic==1.10.7
uvicorn==0.21.1
httpx==0.24.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
PyJWT==2.7.0
